+++ client/src/components/App/App.test.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { mount } from 'enzyme'
 import renderer from 'react-test-renderer'
-import App from './App'
+import App from '.'
 
 describe('<App />', () => {
   const app = mount(<App />)
